requirements.txt
python-barcode
setuptools
pillow
reportlab
